Implement filterBlocksToShow
